Menu bugfixes and ShopButton Interaction
I fixed some bg and menu animation bugs, I added that on every menu animation, the elements goes to the direction of the anchor that has set. I also started the Shop Menu and added the interaction to the Character Buttons. A Dropdown is deployed everytime you click it, and you can also close the window. You can't buy the item yet. .-.